% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{calculateZpfactor}
\alias{calculateZpfactor}
\title{\eqn{Z' factor}}
\usage{
calculateZpfactor(positiveControls, negativeControls)
}
\arguments{
\item{positiveControls}{vector of positive controls values. In The Pamgene context it is the left and right bright controls points.}

\item{negativeControls}{vector of negative controls values. In The Pamgene context it is background points.}
}
\value{
Return \eqn{Z' factor} value.
}
\description{
The \eqn{Z' factor} is a measure of the assay quality, showing the separation between the distribution of the positive and negative controls. one use the background as negative control and the bright left and right points as positive controls.
}
