% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{check_registration}
\alias{check_registration}
\title{Plot PamChip images alignment}
\usage{
check_registration(source, target, nLevels = 4, maxIterations = 5,
  useBlockPercentage = 50)
}
\arguments{
\item{source}{the Pamgene image to be aligned.}

\item{target}{The Pamgene image to align on.}

\item{nLevels}{\code{\link[RNiftyReg]{niftyreg.nonlinear}} parameter (default 4).}

\item{maxIterations}{maxIterations \code{\link[RNiftyReg]{niftyreg.nonlinear}} parameter (default 5).}

\item{useBlockPercentage}{\code{\link[RNiftyReg]{niftyreg.nonlinear}} parameter (default 50).}
}
\value{
Returns the target image with on top the edges of the source image. The red edges should not be shifted.
}
\description{
Plot the registration the source on the target image to check that the alignment was done correctly.
}
