% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{extract_signal}
\alias{extract_signal}
\title{Extract raw signal from PamChip image}
\usage{
extract_signal(centers.coords, img, radius = 5, parallel = FALSE)
}
\arguments{
\item{centers.coords}{data frame of coordinates of centers obtained from the \code{find_centers} function.}

\item{img}{Pamgene image.}

\item{radius}{the number of pixels to take from the center to build the circle (default 5).}

\item{parallel}{Parallelize to process to fasten the output.}
}
\value{
Returns a data frame with the basic statistics values of all pixels in the circle (mean, median, sum ad sd) of each peptides, left ans right controls and 8 background points.
}
\description{
function to extract raw signal of each peptide, left and right controls and 8 background points.
}
