% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_centers.R
\name{find_centers}
\alias{find_centers}
\title{Find blobs centers}
\usage{
find_centers(image, thresh = "99.7\%")
}
\arguments{
\item{image}{PamChip image.}

\item{thresh}{threshold the image to get only the pixels with the highest values. either "auto" or or percentage string (default "99.7%").}
}
\value{
Returns a data frame with the \code{x} and \code{y} coordinates of the centers, thier row and column number.
}
\description{
Find in the PamChip image the center of each peptide, left and right controls and 8 background points. Use the \code{\link{plot_centers}} function to check that the centers were detected correcly and adapt the threshold with the \code{thresh} parameter.
}
