% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamgeneObject_class.R
\name{pamgeneObject}
\alias{pamgeneObject}
\title{Pamgene object from the extracted signal .txt file}
\usage{
pamgeneObject(file, data, chipID, well, exptime, cycle)
}
\arguments{
\item{file}{the extracted signal .txt file name.}
}
\value{
Takes as input the .txt file and return an object containing the data for each experiment with the corresponding metadata: chipID, well, exposure time and cycle
}
\description{
Takes as input the .txt file and return an object containing the data for each experiment with the corresponding metadata: chipID, well, exposure time and cycle
}
