% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_centers.R
\name{plot_centers}
\alias{plot_centers}
\title{Plot centers}
\usage{
plot_centers(image, centers_coords)
}
\arguments{
\item{image}{PamChip image.}

\item{centers_coords}{data frame of coordinates of centers obtained from the \code{\link{find_centers}} function.}
}
\value{
Returns the PamChip image with the centers of each peptides, left ans right controls and 8 background points.
}
\description{
Plot the PamChip image with the detecte centers from the \code{\link{find_centers}} function. This function is used to check that the centers are detected correcly.
}
