% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{plot_circles}
\alias{plot_circles}
\title{Plot circles on PamChip image}
\usage{
plot_circles(centers.coords, img, radius = 5, parallel = TRUE)
}
\arguments{
\item{centers.coords}{data frame of coordinates of centers obtained from the \code{link{find_centers}} function.}

\item{img}{PamChip image.}

\item{radius}{the number of pixels to take from the center to build the circle (default 5).}

\item{parallel}{Logical. If TRUE parallelize to process to fasten the output (default TRUE).}
}
\value{
Returns the PamChip image with the circles of areas of each peptides, left ans right controls and 8 background points.
}
\description{
Plot the circles on the PamChip image of the areas from wich the signal will be taken with the \code{\link{extract_signal}} function. this function is used to optimize the radius to be used and check that the circles are inside the blobs.
}
