% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{plot_image_from_signal}
\alias{plot_image_from_signal}
\title{Plot extracted signal as a PamChip image}
\usage{
plot_image_from_signal(signal)
}
\arguments{
\item{signal}{data frame from the \code{\link{extract_signal}} function output.}
}
\value{
Returns the Pamgene image with the circles of areas of each peptides, left ans right controls and 8 background points. The resulting plot should looks like the original PanChip image.
}
\description{
Plot the extracted signal from \code{\link{extract_signal}} function in the form of the original PamChip image. Used check that the signal was extracted correctly. The resulting plot should looks like the original PamChip image.
}
