% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{registerAllImages}
\alias{registerAllImages}
\title{Align all images in a directory on a target image and extract summary  statistics from each aligned image.}
\usage{
Register (align) all the images in a directory on the chosen target.
}
\arguments{
\item{inDirectory}{The directory of the Pamgene images to be aligned.}

\item{outDirectory}{The output directoy to put the extracted signal and registered image on.}

\item{target}{The Pamgene image to align on.}

\item{nLevels}{RNiftyReg parameter.}

\item{maxIterations}{RNiftyReg parameter.}

\item{useBlockPercentage}{RNiftyReg parameter.}

\item{centers.coords}{data frame of coordinates of the target image of centers obtained from the \code{find_centers} function.}

\item{radius}{the number of pixels to take from the center to build the circle (default 5).}

\item{parallel.sig.extract}{if TRUE the signal extraction process will be parallelized (default FALSE).}

\item{pamgene.layout.file}{merge the Pamgene STK or PTK layout with the extracred values.}

\item{parallel.registration}{if TRUE the registration process will be parallelized.}

\item{extract.signal}{if TRUE the basic statistics values of each peptides, left ans right controls and 8 background points will be extracted. if FALSE only the aligned (registered) image will be returned.}
}
\value{
Returns for each image a .txt file with the extracted summary statistics for each peptide, left and right controls and 8 background points.
}
