% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{regiter_one_image}
\alias{regiter_one_image}
\title{Align one image on a target image an extract the signal form it}
\usage{
regiter_one_image(source, target, nLevels = 4, maxIterations = 5,
  useBlockPercentage = 50, extract.signal = TRUE, centers.coords,
  radius = 5, parallel.sig.extract = TRUE, pamgene.layout.file)
}
\arguments{
\item{source}{the PamChip image to be aligned.}

\item{target}{The PamChip image to align on.}

\item{nLevels}{\code{\link[RNiftyReg]{niftyreg.nonlinear}} parameter (default 4).}

\item{maxIterations}{\code{\link[RNiftyReg]{niftyreg.nonlinear}} parameter (default 5).}

\item{useBlockPercentage}{\code{\link[RNiftyReg]{niftyreg.nonlinear}} (parameter default 50).}

\item{extract.signal}{if TRUE the basic statistics values of each peptides, left ans right controls and 8 background points will be extracted. if FALSE only the aligned (registered) image will be returned.}

\item{centers.coords}{data frame of coordinates of centers of the target image obtained from the \code{\link{find_centers}} function.}

\item{radius}{the number of pixels to take from the center to build the circle (default 5).}

\item{parallel.sig.extract}{if TRUE the process will be parallelized (default TRUE).}

\item{pamgene.layout.file}{STK or PTK PamChip layout.}
}
\value{
Returns a list with the registered image an the a data frame of extracted signal from it if the \code{extract.signal} parameter is TRUE. Only the registered image otherwise.
}
\description{
Uses the \code{\link{[RNiftyReg]}} package to align (register) the source image on the chosen target then extract the signal from it. the goal is to use one target image for which one extracted the coordinates of centers with the \code{\link{find_centers}} function and use this image as a reference on which one will align the pther images and extract the signal rom the same coordinates
}
