% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{ZpFactor_forAll}
\alias{ZpFactor_forAll}
\title{\eqn{Z' factor} for all samples in a data frame}
\usage{
ZpFactor_forAll(rawData)
}
\arguments{
\item{rawData}{take raw data files as obtained from \code{\link{mergeExperiments}} function.}
}
\value{
Returns the same input data frame with an added column of calculated \eqn{Z' factors} for each row.
}
\description{
Calculate the \eqn{Z' factor} for all PamChip samples (rows) in a data frame. The \eqn{Z' factor} is a measure of the assay quality, showing the separation between the distribution of the positive and negative controls. one use the background as negative control and the bright left and right points as positive controls.
}
